backend/requirements.txt
fastapi
uvicorn[standard]
pymysql
SQLAlchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
requests
email-validator
stripe
razorpay